backend/tests/infrastructure/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-cov==4.1.0
pymongo==4.6.1
redis==5.0.1
kafka-python==2.0.2
elasticsearch==8.11.1
pandas==2.1.4
python-dotenv==1.0.0
docker==7.0.0
requests==2.31.0
